refactor: move supported locales to constants.ts
- Create new constants.ts file with SUPPORTED_LOCALES array and type
- Remove supportedLocales from hubby.d.ts to avoid duplication
- Update imports in partner.d.ts and booking.d.ts to use new location
- Export constants from package root index.d.ts
