+++ index.d.ts
@@ -10,3 +10,4 @@ export * from "./src/esim";
 export * from "./src/api";
 export * from "./src/payment";
 export * from "./src/apiLogs";
+export * from "./src/constants";
